feat: separate mongoclient implementation for scopes: custom_builtins and user roles and bindings retrieve
